+++ src/app/problems/_services/difficulty.service.spec.ts
@@ -4,7 +4,7 @@ import {DifficultyService} from './difficulty.service';
 import {TestModule} from '@test/test.module';
 import {ApiService} from "@app/_services/api.service";
 import {HttpTestingController} from "@angular/common/http/testing";
-import {MOCK_DIFFICULTIES} from "@app/problems/_tests/mock";
+import {MOCK_DIFFICULTIES} from "@app/problems/_test/mock";
 
 describe('DifficultyService', () => {
     let difficultyService: DifficultyService;
