+++ src/app/problems/_services/submission.service.spec.ts
@@ -4,7 +4,7 @@ import {SubmissionService} from './submission.service';
 import {TestModule} from '@test/test.module';
 import {ApiService} from "@app/_services/api.service";
 import {HttpTestingController} from "@angular/common/http/testing";
-import {MOCK_QUESTIONS, MOCK_SUBMISSIONS} from "@app/problems/_tests/mock";
+import {MOCK_QUESTIONS, MOCK_SUBMISSIONS} from "@app/problems/_test/mock";
 
 describe('SubmissionService', () => {
     let submissionService: SubmissionService;
