+++ src/app/problems/_test/problem-create/parsons-create-snippet.component.spec.ts
@@ -1,26 +1,26 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import {JavaCreateSnippetComponent} from '../../problem-create/java-create-snippet/java-create-snippet.component';
+import {ParsonsCreateSnippetComponent} from '../../problem-create/parsons-create-snippet/parsons-create-snippet.component';
 import {TestModule} from '@test/test.module';
 import {CkEditorComponent} from "@app/problems/ck-editor/ck-editor.component";
 import {JsonEditorComponent} from "@app/problems/json-editor/json-editor.component";
 import {CKEditorModule} from "@ckeditor/ckeditor5-angular";
+import {ReactiveFormsModule} from "@angular/forms";
 import {CategoryService} from "@app/_services/api/category.service";
 import {CategoryServiceMock} from "@test/category.service.mock";
 import {CourseService} from "@app/_services/api/course/course.service";
 import {CourseServiceMock} from "@test/course.service.mock";
 import {QuestionService} from "@app/problems/_services/question.service";
-import {QuestionServiceMock} from "@app/problems/_tests/question.service.mock";
-import {ReactiveFormsModule} from "@angular/forms";
+import {QuestionServiceMock} from "@app/problems/_test/question.service.mock";
 import {MOCK_COURSES} from "@test/mock";
 
-describe('JavaCreateSnippetComponent', () => {
-    let component: JavaCreateSnippetComponent;
-    let fixture: ComponentFixture<JavaCreateSnippetComponent>;
+describe('ParsonsCreateSnippetComponent', () => {
+    let component: ParsonsCreateSnippetComponent;
+    let fixture: ComponentFixture<ParsonsCreateSnippetComponent>;
 
     beforeEach(async () => {
         await TestBed.configureTestingModule({
-            declarations: [JavaCreateSnippetComponent, CkEditorComponent, JsonEditorComponent],
+            declarations: [ParsonsCreateSnippetComponent, CkEditorComponent, JsonEditorComponent],
             imports: [TestModule, CKEditorModule, ReactiveFormsModule],
             providers: [
                 {provide: CategoryService, useClass: CategoryServiceMock},
@@ -31,7 +31,7 @@ describe('JavaCreateSnippetComponent', () => {
     });
 
     beforeEach(() => {
-        fixture = TestBed.createComponent(JavaCreateSnippetComponent);
+        fixture = TestBed.createComponent(ParsonsCreateSnippetComponent);
         component = fixture.componentInstance;
         fixture.detectChanges();
     });
@@ -45,8 +45,9 @@ describe('JavaCreateSnippetComponent', () => {
         expect(component.events).toEqual(MOCK_COURSES[0].events);
     });
 
-    it('java create question', () => {
+    it('parsons create question', () => {
         component.formGroup.controls['title'].setValue('Test Title');
+        component.formGroup.controls['lines'].setValue('Test Line');
         component.onSubmit();
 
         // The formGroup is reset upon successful submission.
