+++ src/app/problems/_test/problem-edit/java-edit-snippet.component.spec.ts
@@ -2,7 +2,7 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {JavaEditSnippetComponent} from '../../problem-edit/java-edit-snippet/java-edit-snippet.component';
 import {TestModule} from '@test/test.module';
-import {MOCK_MCQ_QUESTION} from '@app/problems/_tests/mock';
+import {MOCK_MCQ_QUESTION} from '@app/problems/_test/mock';
 
 describe('JavaEditSnippetComponent', () => {
     let component: JavaEditSnippetComponent;
