+++ src/app/problems/_test/problem-set/problem-set.component.spec.ts
@@ -5,7 +5,7 @@ import {TestModule} from '@test/test.module';
 import {CategoryService} from "@app/_services/api/category.service";
 import {CategoryServiceMock} from "@test/category.service.mock";
 import {DifficultyService} from "@app/problems/_services/difficulty.service";
-import {DifficultyServiceMock} from "@app/problems/_tests/difficulty.service.mock";
+import {DifficultyServiceMock} from "@app/problems/_test/difficulty.service.mock";
 
 describe('ProblemSetComponent', () => {
     let component: ProblemSetComponent;
