+++ src/app/problems/_test/problem-view/parsons-view-snippet.component.spec.ts
@@ -3,9 +3,9 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {ParsonsViewSnippetComponent} from '../../problem-view/parsons-view-snippet/parsons-view-snippet.component';
 import {TestModule} from '@test/test.module';
 import {DragulaModule} from "ng2-dragula";
-import {MOCK_UQJ_4} from "@app/problems/_tests/mock";
+import {MOCK_UQJ_4} from "@app/problems/_test/mock";
 import {SubmissionService} from "@app/problems/_services/submission.service";
-import {SubmissionServiceMock} from "@app/problems/_tests/submission.service.mock";
+import {SubmissionServiceMock} from "@app/problems/_test/submission.service.mock";
 
 describe('ParsonsViewSnippetComponent', () => {
     let component: ParsonsViewSnippetComponent;
